@@ -8,52 +8,46 @@
     "build": "ng build --prod",
     "test": "ng test --watch=false --browsers=ChromeHeadlessCustom",
     "lint": "ng lint",
-    "e2e": "ng e2e",
-    "extract": "ngx-translate-extract -m _ --input ./src --output ./src/assets/i18n/de.json ./src/assets/i18n/en.json --clean --sort --format namespaced-json"
+    "e2e": "ng e2e"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "7.2.15",
-    "@angular/cdk": "^7.1.0",
-    "@angular/common": "7.2.15",
-    "@angular/compiler": "7.2.15",
-    "@angular/core": "7.2.15",
-    "@angular/flex-layout": "^7.0.0-beta.19",
-    "@angular/forms": "7.2.15",
-    "@angular/http": "7.2.15",
-    "@angular/material": "^7.1.0",
-    "@angular/platform-browser": "7.2.15",
-    "@angular/platform-browser-dynamic": "7.2.15",
-    "@angular/router": "7.2.15",
-    "@angular/service-worker": "7.2.15",
-    "@ngx-translate/core": "^10.0.2",
-    "@ngx-translate/http-loader": "^3.0.1",
+    "@angular/animations": "9.0.0",
+    "@angular/cdk": "^9.0.0",
+    "@angular/common": "9.0.0",
+    "@angular/compiler": "9.0.0",
+    "@angular/core": "9.0.0",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "9.0.0",
+    "@angular/material": "^9.0.0",
+    "@angular/platform-browser": "9.0.0",
+    "@angular/platform-browser-dynamic": "9.0.0",
+    "@angular/router": "9.0.0",
+    "@angular/service-worker": "9.0.0",
+    "@ngx-translate/core": "^12.0.0",
+    "@ngx-translate/http-loader": "^4.0.0",
     "@stomp/ng2-stompjs": "^7.2.0",
-    "chart.js": "^2.7.3",
+    "chart.js": "^2.9.3",
     "core-js": "^2.5.7",
-    "hammerjs": "^2.0.8",
     "is-docker": "^1.1.0",
     "material-design-icons": "^3.0.1",
-    "ng2-charts": "^1.6.0",
-    "ngx-markdown": "^1.6.0",
-    "ngx-qrcode2": "0.0.9",
-    "rxjs": "^6.5.2",
-    "rxjs-compat": "^6.2.2",
-    "tslib": "^1.9.0",
+    "ng2-charts": "^2.3.0",
+    "ngx-markdown": "^8.2.2",
+    "rxjs": "^6.5.4",
+    "tslib": "^1.10.0",
     "tslint-eslint-rules": "^5.3.1",
     "typescript-map": "0.0.7",
-    "zone.js": "^0.8.26"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.7",
-    "@angular/cli": "7.3.9",
-    "@angular/compiler-cli": "^7.2.11",
-    "@angular/language-service": "7.2.15",
-    "@biesbjerg/ngx-translate-extract": "^2.3.4",
+    "@angular-devkit/build-angular": "~0.900.0",
+    "@angular/cli": "9.0.0",
+    "@angular/compiler-cli": "^9.0.0",
+    "@angular/language-service": "9.0.0",
     "@types/jasmine": "^2.8.8",
     "@types/jasminewd2": "~2.0.2",
-    "@types/node": "^6.0.116",
-    "codelyzer": "^4.5.0",
+    "@types/node": "^12.11.1",
+    "codelyzer": "^5.1.2",
     "jasmine-core": "~2.8.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^2.0.5",
@@ -64,6 +58,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~4.1.0",
     "tslint": "^5.9.1",
-    "typescript": "3.2.4"
+    "typescript": "3.7.5"
   }
 }
