@@ -2,7 +2,8 @@ import { NgModule } from '@angular/core';
 import { ArsComponent } from './ars.component';
 import { ArsSliderDirective } from './components/io/slider/ars-slider.directive';
 import { ArsSliderCombComponent } from './components/io/slider/ars-slider-comb/ars-slider-comb.component';
-import { MatButtonModule, MatIconModule } from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
 import { FullScreenOverlayComponent } from './components/layout/base/full-screen-overlay/full-screen-overlay.component';
 import { RowComponent } from './components/layout/frame/row/row.component';
 import { ColComponent } from './components/layout/frame/col/col.component';
