@@ -1,6 +1,6 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { ContentChoiceCreatorComponent, DisplayAnswer } from '../../content-choice-creator/content-choice-creator.component';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { EventService } from '../../../../services/util/event.service';
 
 
