@@ -2,7 +2,8 @@ import { Component, Inject, OnChanges, OnInit, SimpleChanges } from '@angular/co
 import { AuthenticationService } from '../../../../services/http/authentication.service';
 import { Router } from '@angular/router';
 import { NotificationService } from '../../../../services/util/notification.service';
-import { ErrorStateMatcher, MAT_DIALOG_DATA, MatDialog } from '@angular/material';
+import { ErrorStateMatcher } from '@angular/material/core';
+import { MAT_DIALOG_DATA, MatDialog } from '@angular/material/dialog';
 import { FormControl, FormGroupDirective, NgForm, Validators } from '@angular/forms';
 import { UserRole } from '../../../../models/user-roles.enum';
 import { TranslateService } from '@ngx-translate/core';
