@@ -1,5 +1,5 @@
 import { AfterContentInit, Component, OnDestroy, OnInit, Renderer2 } from '@angular/core';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { TranslateService } from '@ngx-translate/core';
 import { LanguageService } from '../../../services/util/language.service';
 import { RoomCreateComponent } from '../../shared/_dialogs/room-create/room-create.component';
