@@ -1,5 +1,5 @@
 import { Component, Inject, OnInit } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { RoomEditComponent } from '../../../creator/_dialogs/room-edit/room-edit.component';
 import { RoomService } from '../../../../services/http/room.service';
 import { Room } from '../../../../models/room';
