@@ -6,7 +6,7 @@ import { User } from '../../../models/user';
 import { UserRole } from '../../../models/user-roles.enum';
 import { Location } from '@angular/common';
 import { TranslateService } from '@ngx-translate/core';
-import { MatDialog } from '@angular/material';
+import { MatDialog } from '@angular/material/dialog';
 import { LoginComponent } from '../../home/_dialogs/login/login.component';
 import { DeleteAccountComponent } from '../_dialogs/delete-account/delete-account.component';
 import { UserService } from '../../../services/http/user.service';
@@ -15,7 +15,6 @@ import { KeyboardUtils } from '../../../utils/keyboard';
 import { KeyboardKey } from '../../../utils/keyboard/keys';
 import { UserBonusTokenComponent } from '../_dialogs/user-bonus-token/user-bonus-token.component';
 import { RemindOfTokensComponent } from '../_dialogs/remind-of-tokens/remind-of-tokens.component';
-import { QrCodeDialogComponent } from '../_dialogs/qr-code-dialog/qr-code-dialog.component';
 import { BonusTokenService } from '../../../services/http/bonus-token.service';
 
 @Component({
@@ -219,15 +218,15 @@ export class HeaderComponent implements OnInit {
   }
 
   public showQRDialog() {
-    document.body.requestFullscreen();
-    const dialogRef = this.dialog.open(QrCodeDialogComponent, {
-      panelClass: 'screenDialog'
-    });
-    const qrDialog: QrCodeDialogComponent = dialogRef.componentInstance;
-    qrDialog.setQRCode(this.getQRCode());
-    dialogRef.afterClosed().subscribe(res => {
-      document.exitFullscreen();
-    });
+    // document.body.requestFullscreen();
+    // const dialogRef = this.dialog.open(QrCodeDialogComponent, {
+    //   panelClass: 'screenDialog'
+    // });
+    // const qrDialog: QrCodeDialogComponent = dialogRef.componentInstance;
+    // qrDialog.setQRCode(this.getQRCode());
+    // dialogRef.afterClosed().subscribe(res => {
+    //   document.exitFullscreen();
+    // });
   }
 
 }
